refactor bench utils and other comments
