+++ lib/kv-router/benches/kv_indexer_bench.rs
@@ -13,11 +13,9 @@
 
 use clap::{Args, Parser, Subcommand, ValueEnum};
 use dynamo_kv_router::{
+    bench_utils::{LatencyStats, SequenceData, generate_sequences, median},
     indexer::{KvIndexer, KvIndexerInterface, KvIndexerMetrics, KvIndexerSharded},
-    protocols::{
-        ExternalSequenceBlockHash, KvCacheEvent, KvCacheEventData, KvCacheRemoveData,
-        KvCacheStoreData, KvCacheStoredBlockData, LocalBlockHash, RouterEvent, WorkerId,
-    },
+    protocols::{LocalBlockHash, RouterEvent},
 };
 use std::sync::Arc;
 use std::sync::atomic::{AtomicU64, Ordering};
@@ -150,165 +148,6 @@ struct StressArgs {
     /// Number of shards for sharded indexer
     #[arg(long, default_value = "4")]
     num_shards: usize,
-}
-
-// ============================================================================
-// Shared Data Structures
-// ============================================================================
-
-/// Pre-generated sequence data for benchmarking
-#[derive(Clone)]
-struct SequenceData {
-    worker_id: WorkerId,
-    local_hashes: Vec<LocalBlockHash>,
-    external_hashes: Vec<ExternalSequenceBlockHash>,
-}
-
-impl SequenceData {
-    fn new(seq_id: u64, worker_id: WorkerId, depth: usize) -> Self {
-        let local_hashes: Vec<LocalBlockHash> = (0..depth)
-            .map(|block_idx| LocalBlockHash((seq_id << 32) | (block_idx as u64)))
-            .collect();
-
-        let external_hashes: Vec<ExternalSequenceBlockHash> = (0..depth)
-            .map(|block_idx| ExternalSequenceBlockHash((seq_id << 32) | (block_idx as u64)))
-            .collect();
-
-        Self {
-            worker_id,
-            local_hashes,
-            external_hashes,
-        }
-    }
-
-    fn to_store_event(&self, event_id: u64) -> RouterEvent {
-        RouterEvent {
-            worker_id: self.worker_id,
-            event: KvCacheEvent {
-                event_id,
-                data: KvCacheEventData::Stored(KvCacheStoreData {
-                    parent_hash: None,
-                    blocks: self
-                        .local_hashes
-                        .iter()
-                        .zip(self.external_hashes.iter())
-                        .map(|(local, ext)| KvCacheStoredBlockData {
-                            tokens_hash: *local,
-                            block_hash: *ext,
-                            mm_extra_info: None,
-                        })
-                        .collect(),
-                }),
-                dp_rank: 0,
-            },
-        }
-    }
-
-    fn to_remove_event(&self, event_id: u64) -> RouterEvent {
-        RouterEvent {
-            worker_id: self.worker_id,
-            event: KvCacheEvent {
-                event_id,
-                data: KvCacheEventData::Removed(KvCacheRemoveData {
-                    block_hashes: self.external_hashes.clone(),
-                }),
-                dp_rank: 0,
-            },
-        }
-    }
-}
-
-/// Generate sequences with shared prefix prompts
-fn generate_sequences(
-    num_sequences: usize,
-    depth: usize,
-    num_workers: usize,
-    prefix_ratio: f64,
-    num_prefix_groups: usize,
-    seed: u64,
-) -> Vec<SequenceData> {
-    use rand::{Rng, SeedableRng, rngs::StdRng};
-
-    let mut sequences = Vec::with_capacity(num_sequences);
-    let prefix_length: usize = (depth as f64 * prefix_ratio).round() as usize;
-    let mut rng: StdRng = StdRng::seed_from_u64(seed);
-
-    for seq_id in 0..num_sequences {
-        let worker_id = (seq_id % num_workers) as WorkerId;
-        let mut seq = SequenceData::new(seq_id as u64, worker_id, depth);
-
-        if num_prefix_groups > 0 && prefix_length > 0 {
-            let group_id = rng.random_range(0..num_prefix_groups);
-            for i in 0..prefix_length {
-                seq.local_hashes[i] =
-                    LocalBlockHash(0xDEAD_BEEF_0000_0000 | ((group_id as u64) << 32) | (i as u64));
-            }
-        }
-
-        sequences.push(seq);
-    }
-
-    sequences
-}
-
-/// Statistics for latency measurements
-#[derive(Debug, Clone)]
-struct LatencyStats {
-    min: Duration,
-    max: Duration,
-    avg: Duration,
-    p50: Duration,
-    p95: Duration,
-    p99: Duration,
-    throughput_ops_sec: f64,
-}
-
-impl LatencyStats {
-    fn from_durations(mut durations: Vec<Duration>) -> Option<Self> {
-        if durations.is_empty() {
-            return None;
-        }
-
-        durations.sort();
-        let n = durations.len();
-        let total: Duration = durations.iter().sum();
-        let avg = total / n as u32;
-
-        Some(Self {
-            min: durations[0],
-            max: durations[n - 1],
-            avg,
-            p50: durations[n / 2],
-            p95: durations[n * 95 / 100],
-            p99: durations[n * 99 / 100],
-            throughput_ops_sec: n as f64 / total.as_secs_f64(),
-        })
-    }
-
-    fn print(&self, operation: &str, blocks_per_op: usize) {
-        println!("\n{} Latency Statistics:", operation);
-        println!("  min:  {:>12?}", self.min);
-        println!("  avg:  {:>12?}", self.avg);
-        println!("  p50:  {:>12?}", self.p50);
-        println!("  p95:  {:>12?}", self.p95);
-        println!("  p99:  {:>12?}", self.p99);
-        println!("  max:  {:>12?}", self.max);
-        println!("  throughput: {:.2} ops/sec", self.throughput_ops_sec);
-        println!(
-            "  throughput: {:.2} blocks/sec",
-            self.throughput_ops_sec * blocks_per_op as f64
-        );
-    }
-}
-
-/// Compute median of durations
-fn median(durations: &[Duration]) -> Duration {
-    if durations.is_empty() {
-        return Duration::ZERO;
-    }
-    let mut sorted = durations.to_vec();
-    sorted.sort();
-    sorted[sorted.len() / 2]
 }
 
 // ============================================================================
@@ -844,6 +683,7 @@ async fn run_microbench_mode(args: MicrobenchArgs) {
         args.prefix_prompt_ratio,
         args.num_prefix_prompts,
         args.common.seed,
+        false,
     );
     let sequences = &all_sequences[..num_sequences];
     let extra_sequences = &all_sequences[num_sequences..];
@@ -1361,6 +1201,7 @@ async fn run_stress_mode(args: StressArgs) {
         args.prefix_share_ratio,
         1, // Single prefix group for stress test
         args.common.seed,
+        false, // use_cumulative_hash
     );
 
     let metrics = Arc::new(KvIndexerMetrics::new_unregistered());
